feat: add expose to docgen-cli
